requirements.txt
numpy>=1.25.4
pandas>=2.2.2
matplotlib>=3.8.4
seaborn>=0.13.2
scikit-learn>=1.4.2
tensorflow-macos>=2.15.0 tensorflow-metal
xgboost>=2.0.3
optuna>=3.6.0
tqdm>=4.66.4